Finishes packet handling.
Replaces pen draw with arrow draw.
Edits the way draw packets are deserialized and invert x1 x2 or y1 y2 depending on which value is the lowest.
Adds a way to synchronize every client with the server when they connect. So each time a client connects, it requests a draw synchronization.